# docs/api/openapiv1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AzureBank.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/{azureTag}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Look up a recipient by AzureTag",
        "description": "Look up a single user by their EXACT AzureTag, to confirm a transfer recipient.\r\nReturns the masked display name (e.g. \"Vladislav A.\") for confirmation. This is an\r\nexact-match confirmation oracle by design — there is deliberately no substring/prefix\r\ndirectory search, which would let an authenticated user harvest the customer list\r\n(ADR-0014; the Zelle/Cash App model).",
        "parameters": [
          {
            "name": "azureTag",
            "in": "path",
            "description": "Full AzureTag to look up (3-20 chars, AzureTag charset).",
            "required": true,
            "schema": {
              "maxLength": 20,
              "minLength": 3,
              "pattern": "^[a-z][a-z0-9_]{2,19}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfRecipientLookupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication failed or is missing. The body is a ProblemDetails whose errorCode names the reason (e.g. AUTH_TOKEN_MISSING, AUTH_TOKEN_INVALID).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - authenticated, but not permitted to reach this resource (errorCode: ACCESS_DENIED).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/azuretag": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Rename my AzureTag",
        "description": "Renames the caller's own public AzureTag handle (ADR-0015). The handle is decoupled\r\nfrom the login identity (UserName is the immutable user id), so this is a plain update.\r\nThe bearer token still carries the old handle in its azure_tag claim until it is\r\nrefreshed on next login.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAzureTagRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAzureTagRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAzureTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfUpdateAzureTagResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - the requested AzureTag is already taken (errorCode AZURE_TAG_TAKEN).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "A URI reference identifying the problem type"
                    },
                    "title": {
                      "type": "string",
                      "description": "A short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "description": "The HTTP status code (409)"
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation of the failure"
                    },
                    "errorCode": {
                      "type": "string",
                      "description": "Machine-readable error code (AZURE_TAG_TAKEN)"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Request trace identifier for debugging"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication failed or is missing. The body is a ProblemDetails whose errorCode names the reason (e.g. AUTH_TOKEN_MISSING, AUTH_TOKEN_INVALID).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - authenticated, but not permitted to reach this resource (errorCode: ACCESS_DENIED).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponseOfRecipientLookupResponse": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RecipientLookupResponse"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfUpdateAzureTagResponse": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UpdateAzureTagResponse"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorCode": {
            "type": "string",
            "description": "Stable, machine-readable identifier for the failure — the value clients branch on. Absent only on validation failures, which carry an 'errors' dictionary instead."
          },
          "traceId": {
            "type": "string",
            "description": "Correlates this response with the server-side log entry. Quote it in a bug report."
          }
        }
      },
      "RecipientLookupResponse": {
        "required": [
          "azureTag",
          "displayName",
          "exists"
        ],
        "type": "object",
        "properties": {
          "azureTag": {
            "type": "string",
            "description": "The user's AzureTag"
          },
          "displayName": {
            "type": "string",
            "description": "Masked display name for privacy (e.g., \"John D.\")"
          },
          "exists": {
            "type": "boolean",
            "description": "Whether the user exists in the system"
          }
        },
        "description": "Response for recipient lookup by AzureTag.\r\nUsed for transfer recipient verification."
      },
...
      "UpdateAzureTagRequest": {
        "required": [
          "azureTag"
        ],
        "type": "object",
        "properties": {
          "azureTag": {
            "maxLength": 20,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_]{2,19}$",
            "type": "string"
          }
        },
        "description": "Request to rename the caller's own public AzureTag handle (ADR-0015)."
      },
      "UpdateAzureTagResponse": {
        "required": [
          "azureTag"
        ],
        "type": "object",
        "properties": {
          "azureTag": {
            "type": "string"
          }
        },
        "description": "The caller's AzureTag after a successful rename (ADR-0015)."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT token. Get a token by calling POST /api/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}